+++ nym-vpn-core/crates/nym-vpnd/src/service/config/config_manager.rs
@@ -82,6 +82,7 @@ impl VpnServiceConfigManager {
         &self.config
     }
 
+    #[cfg(test)]
     pub async fn set_config(&mut self, config: nym_vpn_lib_types::VpnServiceConfig) {
         if self.config != config {
             self.config = config;